window length should be integer

+++ Sources/CP3_DSP_ObjC/KarplusVoice.cpp
@@ -75,7 +75,7 @@ double KarplusVoice::process() {
         int index = (mBufferPos + i) % mPeriod;
         sum += mBuffer[index];
     }
-    mBuffer[mBufferPos] = sum / (double)mWindowLength;
+    mBuffer[mBufferPos] = sum / mWindowLength;
     
     return mBuffer[mBufferPos];
 }
